Allow shadowed fields to be set without aliasing
Ambiguities may arise when performing a query across multiple tables with columns of the same name. Previously, it was advised that ambiguous fields should be aliased to provide a unique name. This modification modifies the fetch process to permit assigning fields of the same name.

  // BookAuthor
  type BookAuthor struct {
    // Both Author and Book have and ID column, we need this extra field to tell
    // the difference between the ID of the book and the ID of the author.
    BookID int `db:"book_id"`

    Author `db:",inline"`
    Book   `db:",inline"`
  }

may now be represented as

  // BookAuthor
  type BookAuthor struct {
    Author `db:",inline"`
    Book   `db:",inline"`
  }

Ambiguities are resolved by precedence. For example, the first `id` column will be assigned to the Author ID field and the second `id` column will be assigned to the Book ID field. Represented in-query as: `SELECT author.*, book.* FROM book, author`.