Allow db columns to be nullable
This update adds a db tag that allows a field to be NULL without having to assign the field type to sql.NullInt64 or similar. If the column is NULL, the Go default value will be used in its place. If the value must be identified as NULL, sql.NullX types should still be used.